+++ glutin/src/api/egl/mod.rs
@@ -439,6 +439,7 @@ impl Context {
                 version,
                 pf_reqs,
                 surface_type,
+                opengl,
                 config_selector,
             )?
         };
@@ -969,6 +970,26 @@ impl<'a> ContextPrototype<'a> {
             }
         };
 
+        if let Some(surface) = surface {
+            // VSync defaults to enabled; disable it if it was not requested.
+            if !self.opengl.vsync {
+                let _guard = MakeCurrentGuard::new(
+                    self.display,
+                    surface,
+                    surface,
+                    context,
+                )
+                .map_err(|err| CreationError::OsError(err))?;
+
+                let egl = EGL.as_ref().unwrap();
+                unsafe {
+                    if egl.SwapInterval(self.display, 0) == ffi::egl::FALSE {
+                        panic!("finish_impl: eglSwapInterval failed: 0x{:x}", egl.GetError());
+                    }
+                }
+            }
+        }
+
         Ok(Context {
             display: self.display,
             context,
@@ -988,6 +1009,7 @@ unsafe fn choose_fbconfig<F>(
     version: Option<(u8, u8)>,
     pf_reqs: &PixelFormatRequirements,
     surface_type: SurfaceType,
+    opengl: &GlAttributes<&Context>,
     mut config_selector: F,
 ) -> Result<(ffi::egl::types::EGLConfig, PixelFormat), CreationError>
 where
@@ -1157,7 +1179,42 @@ where
         ));
     }
 
-    if num_configs == 0 {
+    // We're interested in those configs which allow our desired VSync.
+    let desired_swap_interval = if opengl.vsync {
+        1
+    } else {
+        0
+    };
+
+    let config_ids = config_ids.into_iter().filter(|&config| {
+        let mut min_swap_interval = 0;
+        let res = egl.GetConfigAttrib(
+            display,
+            config,
+            ffi::egl::MIN_SWAP_INTERVAL as ffi::egl::types::EGLint,
+            &mut min_swap_interval,
+        );
+
+        if desired_swap_interval < min_swap_interval {
+            return false;
+        }
+
+        let mut max_swap_interval = 0;
+        let res = egl.GetConfigAttrib(
+            display,
+            config,
+            ffi::egl::MAX_SWAP_INTERVAL as ffi::egl::types::EGLint,
+            &mut max_swap_interval,
+        );
+
+        if desired_swap_interval > max_swap_interval {
+            return false;
+        }
+
+        true
+    }).collect::<Vec<_>>();
+
+    if config_ids.is_empty() {
         return Err(CreationError::NoAvailablePixelFormat);
     }